Default char to unsigned to make it harder to write dangerous code.
In the interest of making it harder to shoot yourself in the foot,
have the compiler treat char as an unsigned type by default to prevent
things like: char c = '0xF0'; if (some_table[c]) ... from exploding
spectacularly.

Also constify some things in notify.c that didn't need to be mutable.