Httpd (#1214)
(From "help http"):
PennMUSH now supports HTTP requests reaching its mush port. It is very low level,
and a little tricky to understand.

When an HTTP request hits the PennMUSH port, PennMUSH invisibly logs in to
the HTTP Handler player (@config http_handler), and executes an
@include me/<method>. e.g: @include me/get.

%0 will be the pathname, e.g: "/", "/path/to", "/foo?bar=baz", etc.

%1 will be the body of the request. If it's json, use json_query to deal with it.
If it's form-encoded, look at help formdecode()

Anything sent to the HTTP Handler player during evaluation of this code is
included in the body sent to the HTTP Client. There is a maximum size of
BUFFER_LEN for the body of the response.