Close open sqlite3 file databases before forking.
Leaving a db open across a fork and then accessing it from the child is
a good way to get a corrupt database. While we don't do anything after a
fork that would use an open db connection, that might change in the
future, and it's better to be safe than sorry anyways. Right now they're
only reopened in the parent. In the future if a child needs access to a
db too it can be set up to automatically reopen as well.